requirements.txt
openai-whisper==20231117
PyQt5==5.15.10
moviepy==1.0.3
jieba==0.42.1
nltk==3.8.1
python-vlc==3.0.20123
requests==2.31.0
